+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Cross-file Quick Fixes for opened files
+-   Show diagnostic name
 
 ## [0.6.0] - 2025-01-05